requirements.txt
# Malicious/Vulnerable Python Dependencies POC
# WARNING: These packages contain known security vulnerabilities

# Known malicious/compromised packages
ctx==0.1.2
colourama==0.3.9
python3-dateutil==2.4.0
jellyfish==0.5.6

# Known vulnerable versions
Django==2.2.0
Flask==0.12.2
requests==2.6.0
Pillow==6.2.0
PyYAML==5.1
Jinja2==2.10.0
urllib3==1.24.1
cryptography==2.6.1
paramiko==2.4.0
SQLAlchemy==1.2.0
